Extract OutputTab parameters to core package
