+++ src/main/java/org/bric/gui/output/OutputTab.java
@@ -1,6 +1,6 @@
 package org.bric.gui.output;
 
-import org.bric.core.output.model.OutputParameters;
+import org.bric.core.model.output.OutputParameters;
 import org.bric.gui.BricUI;
 import org.bric.gui.tabs.ImageEditTab;
 import org.bric.utils.Utils;
